query is saved in connection
