+++ public/controllers/SaveQryModalCtrl.js
@@ -1,0 +1,56 @@
+hotChocolate.controller('SaveQryModalCtrl',
+    function ($scope, $rootScope, $uibModalInstance, $timeout, items, queryList, mdxQuery, saveQuery )
+    {
+        $scope.items = items;
+        $scope.queryList = queryList;
+        console.log($scope.queryList);
+        $scope.mdxQuery = mdxQuery;
+       /*************** What to be done for saving **********/
+       $scope.save = function () {
+           console.log($scope.items);
+           var colArray = $scope.items[0].list.length>0 ? $scope.items[0].list : [];
+               rowArray = $scope.items[1].list.length>0 ? $scope.items[1].list : [];
+               filterArray = $scope.items[2].list.length>0 ? $scope.items[2].list : [];
+           var parameters = {
+             queryName: $scope.newQueryName,
+             userName: "hotChocolate",
+             onColumns: colArray,
+             onRows: rowArray,
+             onFilters: filterArray,
+             queryMDX: $scope.mdxQuery,
+             connectionData: {
+               // xmlaServer: "http://172.23.238.252:8080/pentaho/Xmla?userid=admin&password=password",
+               dataSource: $scope.$root.DataSourceName,
+               catalog: $scope.$root.CatalogName,
+               cube: $scope.$root.CubeName,
+             }
+           };
+           var connId = $rootScope.connId;
+           saveQuery.saveQuery(parameters, connId).success(function(data) {
+             console.log(data);
+             $scope.showModalAlert = true;
+             $timeout(function() {
+               $scope.showModalAlert = false;
+             }, 2000);
+             $scope.querySaveMessage = data.info;
+             if(data.status=="success")
+             {
+               $scope.queryList.push({queryName : $scope.newQueryName});
+             }
+             console.log($scope.queryList);
+             if (data.status === "success"){
+               $timeout(function() {
+                 $uibModalInstance.close($scope.queryList);
+               }, 5000);
+             }
+           });
+       };
+       $scope.close = function(){
+         $uibModalInstance.close($scope.queryList);
+       };
+       $scope.cancel = function () {
+         $uibModalInstance.dismiss('cancel');
+       };
+
+        //  $uibModalInstance.close();
+     });
