+++ routes/queryController.js
@@ -1,5 +1,5 @@
 var express = require('express'),
-    Query = require('../models/queryModel');
+    Connections = require('../models/Connections');
 
 var router = express.Router();
 
@@ -30,36 +30,62 @@ router.post('/new', function(req, res) {
   var parameters = JSON.parse(req.body.parameter);
   console.log(parameters);
   console.log("saving query to db");
-  Query.create({
-    queryName: parameters.queryName,
-    createdBy: parameters.userName,
-    createdOn: Date.now(),
-    modifiedOn: Date.now(),
-    onColumns: parameters.colArray,
-    onRows: parameters.rowArray,
-    onFilters: parameters.filterArray,
-    queryMDX: parameters.queryMDX,
-    connectionData: {
-                    //  xmlaServer: parameters.connectionData.xmlaServer,
-                     dataSource: parameters.connectionData.dataSource,
-                     catalog: parameters.connectionData.catalog,
-                     cube: parameters.connectionData.cube
-                   }
-  }, function(err, query) {
-    if(!err) {
-      console.log("query saved to db: " + query);
-      res.json({status: 'success', info: "Query successfully saved"});
-    } else {
-      console.error(err);
-      if(err.code === 11000) {
-        res.json({status: 'error', info: "Query name already exists"});
-      }
-      else {
-        console.log(err);
-        res.json({status: 'error', info: "Oops! error occcured saving query"});
-      }
-    }
-  });
+  var query = parameters.parameters;
+  query.createdOn = Date.now();
+  query.modifiedOn = Date.now();
+  console.log(query);
+  Connections.findOneAndUpdate(
+    {_id : parameters.connId},
+    {
+      $push : {'savedQueries': query}
+    },function(err,connection){
+        if(!err) {
+          console.log(connection);
+          res.json({status: 'success', info: "Query successfully saved"});
+        } else {
+          // console.error(err);
+          if(err.code === 11000) {
+            res.json({status: 'error', info: "Query name already exists"});
+          }
+          else {
+            // console.log(err);
+            res.json({status: 'error', info: "Oops! error occcured saving query"});
+          }
+        }
+    });
+
+
+
+  // Query.create({
+  //   queryName: parameters.queryName,
+  //   createdBy: parameters.userName,
+  //   createdOn: Date.now(),
+  //   modifiedOn: Date.now(),
+  //   onColumns: parameters.colArray,
+  //   onRows: parameters.rowArray,
+  //   onFilters: parameters.filterArray,
+  //   queryMDX: parameters.queryMDX,
+  //   connectionData: {
+  //                   //  xmlaServer: parameters.connectionData.xmlaServer,
+  //                    dataSource: parameters.connectionData.dataSource,
+  //                    catalog: parameters.connectionData.catalog,
+  //                    cube: parameters.connectionData.cube
+  //                  }
+  // }, function(err, query) {
+  //   if(!err) {
+  //     console.log("query saved to db: " + query);
+  //     res.json({status: 'success', info: "Query successfully saved"});
+  //   } else {
+  //     console.error(err);
+  //     if(err.code === 11000) {
+  //       res.json({status: 'error', info: "Query name already exists"});
+  //     }
+  //     else {
+  //       console.log(err);
+  //       res.json({status: 'error', info: "Oops! error occcured saving query"});
+  //     }
+  //   }
+  // });
 });
 
 // get query
