+++ views/index.ejs
@@ -32,7 +32,11 @@
     <!-- Include ConnectionModelController.js -->
     <script src="controllers/ConnectionModelController.js"></script>
     <!-- Include ConnectionModelController.js -->
+    <script src="controllers/ServerCredentialModalCtrl.js"></script>
+    <!-- Include ConnectionModelController.js -->
     <script src="controllers/queryController.js"></script>
+    <!-- Include ConnectionModelController.js -->
+    <script src="controllers/SaveQryModalCtrl.js"></script>
     <!-- Include gridRender.js -->
     <script src="directives/gridRender.js"></script>
     <!-- Include toggleList.js -->
